tools/png2hpp/requirements.txt
autopep8==1.4.4
Click==7.0
Pillow==6.0.0
pycodestyle==2.5.0
